# docs/architecture/api/openapi.yaml
openapi: 3.0.3
info:
  title: Driftlock API
  version: 1.0.0
  description: |
    Compression-based anomaly detection for any JSON event stream.

    ## Quick Start

    1. **Try without signup**: Use `POST /v1/demo/detect` (10 req/min, 50 events max)
    2. **Sign up**: `POST /v1/onboard/signup` → verify email → get API key
    3. **Detect anomalies**: `POST /v1/detect` with `X-Api-Key` header

    ## Authentication

    - **API Key**: Pass `X-Api-Key: dlk_<uuid>.<secret>` header for detection endpoints
    - **Firebase Auth**: Pass `Authorization: Bearer <firebase-id-token>` for dashboard endpoints

    ## Rate Limits

    - Demo endpoint: 10 requests/minute per IP, max 50 events
    - Authenticated: Based on your plan (see /v1/me/usage)
  contact:
    name: Driftlock Support
    email: support@driftlock.net
    url: https://driftlock.net
  license:
    name: Proprietary
    url: https://driftlock.net/terms
servers:
  - url: https://driftlock.net/api
    description: Production API
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/billing/portal:
    post:
      tags: [Billing]
      summary: Open billing portal
      description: Get a link to the Stripe customer portal to manage subscription.
      operationId: billingPortal
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Portal URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  portal_url:
                    type: string
                    format: uri
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        API key in format `dlk_<uuid>.<secret>`

        Get your key from the dashboard or via email verification.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token from client-side Firebase Auth
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - method_not_allowed
                - conflict
                - gone
                - rate_limit_exceeded
                - internal
                - service_unavailable
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request ID for debugging
            retry_after_seconds:
              type: integer
              description: Seconds to wait (only for rate_limit_exceeded)
...